Auth Check At Splash Screen
