+++ OnTimePro/app/src/main/java/com/elevatelab/ontimepro/WelcomeActivity.java
@@ -17,21 +17,14 @@ public class WelcomeActivity extends AppCompatActivity {
 
 
     private Button registerBtn, loginBtn;
-    private FirebaseAuth mFirebaseAuth;
-    private FirebaseFirestore db;
-    private String instiIdOfCurrent;
+
+
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_welcome);
         initViews();
-        mFirebaseAuth = FirebaseAuth.getInstance();
-        db = FirebaseFirestore.getInstance();
-        if (mFirebaseAuth.getCurrentUser() != null) {
-
-            String userID = mFirebaseAuth.getUid();
-            startIntent(userID);
-        }
 
         registerBtn.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -56,23 +49,6 @@ public class WelcomeActivity extends AppCompatActivity {
         loginBtn = findViewById(R.id.login_btn);
     }
 
-    private void startIntent(String userID)
-    {
-        db.collection("Users")
-                .document(userID)
-                .get()
-                .addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
-                    @Override
-                    public void onSuccess(DocumentSnapshot documentSnapshot) {
-                        instiIdOfCurrent = (String) documentSnapshot.get("instiID");
-                        Log.e("INs",instiIdOfCurrent);
-                        Intent mainActivityIntent = new Intent(WelcomeActivity.this, MainActivity.class);
-                        mainActivityIntent.putExtra("instiCode",instiIdOfCurrent);
-                        startActivity(mainActivityIntent);
-                        finish();
-                    }
-                });
 
-    }
 
 }
